ensured text-classification output is type long

+++ src/transformers/sparse.py
@@ -136,10 +136,11 @@ def export_model(exporter, dataloader, output_dir, num_exported_samples):
         if sess is None:
             forward_args_spec = inspect.getfullargspec(exporter._module.__class__.forward)
             one_sample_input = collections.OrderedDict(
-                [(f, sample_batch[f][0].reshape(1, -1)) for f in forward_args_spec.args if f in sample_batch]
+                [(f, sample_batch[f][0].long().reshape(1, -1)) for f in forward_args_spec.args if f in sample_batch]
             )
 
             try:
+
                 exporter.export_onnx(sample_batch=one_sample_input, convert_qat=True)
                 onnx_file = os.path.join(output_dir, "model.onnx")
             except Exception:
@@ -150,7 +151,7 @@ def export_model(exporter, dataloader, output_dir, num_exported_samples):
         input_names = list(sample_batch.keys())
         output_names = [o.name for o in sess.get_outputs()]
         for input_vals in zip(*sample_batch.values()):
-            input_feed = {k: v.numpy() for k, v in zip(input_names, input_vals)}
+            input_feed = {k: v.long().numpy() for k, v in zip(input_names, input_vals)}
             output_vals = sess.run(output_names, {k: input_feed[k].reshape(1, -1) for k in input_feed})
             output_dict = {name: numpy.squeeze(val) for name, val in zip(output_names, output_vals)}
             file_idx = f"{num_samples}".zfill(4)
